feat(pyo3): AlgorithmSequence API + test fixes + type stubs
- Add compositional AlgorithmSequence API with preset constructors
  (nve, nvt, npt, from_parameters) and sequence manipulation
  (add, insert_after, insert_before, remove, replace)
- Add Simulation.from_sequence() for custom algorithm sequences
- Add algorithm descriptor classes: Forcefield, LeapFrogIntegrator,
  LeapFrogVelocity, LeapFrogPosition, BerendsenThermostat,
  BerendsenBarostat, ShakeConstraints, TemperatureCalculation,
  PressureCalculation, EnergyCalculation, RemoveCOMMotion
- Fix test compilation: integer-to-float casts, f32->f64 suffixes,
  cs6/cs12 fields in LJParameters, IMD block format, nearest_image
  test assertions, bspline loop bound cast
- Update PLAN.md with completed sections and test fix notes

+++ crates/gromos-forces/src/bonded.rs
@@ -3062,7 +3062,7 @@ mod tests {
 
         let n_lambda = 11;
         for i in 0..n_lambda {
-            let lambda = i / (n_lambda - 1);
+            let lambda = i as f64 / (n_lambda - 1) as f64;
             let lambda_ctrl = LambdaController::new().with_lambda(lambda);
             let result = calculate_perturbed_angle_forces(&topo, &conf, &lambda_ctrl);
             println!(
